Fixed bug #35570 (segfault when re-using soap client object)

+++ ext/soap/php_http.c
@@ -17,7 +17,7 @@
   |          Dmitry Stogov <dmitry@zend.com>                             |
   +----------------------------------------------------------------------+
 */
-/* $Id: php_http.c,v 1.77.2.4 2005/12/07 13:56:18 dmitry Exp $ */
+/* $Id: php_http.c,v 1.77.2.5 2005/12/12 09:25:23 dmitry Exp $ */
 
 #include "php_soap.h"
 #include "ext/standard/base64.h"
@@ -32,7 +32,7 @@ static int get_http_headers(php_stream *socketd,char **response, int *out_size T
 
 static int stream_alive(php_stream *stream  TSRMLS_DC)
 {
-	int socket;
+	long socket;
 	char buf;
 
 	/* maybe better to use:
